:recycle: refactor: Fixed the merge with dev
