+++ src/main/java/br/fatec/easycoast/repositories/UserRepository.java
@@ -1,8 +1,8 @@
-package br.fatec.easycoast.repository;
+package br.fatec.easycoast.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.fatec.easycoast.entity.User;
+import br.fatec.easycoast.entities.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 }
